requirements.txt
anthropic
yfinance
requests
python-telegram-bot
feedparser
python-dotenv
